Updating method to extract data (rewriting splitImage function as detectVerticalLines)
Instead of manually splitting the page, I want to write a program that detects where the vertical lines are on the page. For every page, the same vertical lines correspond to the same "cut-off" points, so I can use this to make the later text extraction process a lot easier.